Data is already being save on db

+++ db/migrate/20160804192342_create_sales.rb
@@ -3,11 +3,11 @@ class CreateSales < ActiveRecord::Migration
     create_table :sales do |t|
       t.string :customer
       t.string :description
-      t.decimal :price
+      t.float :price
       t.integer :quantity
       t.string :address
       t.string :supplier
-      t.decimal :total
+      t.float :total
 
       t.timestamps null: false
     end
